Moved every and duration tests from SchedulerSpec to StreamSpec
