+++ core/jvm/src/test/scala/fs2/StreamSpec.scala
@@ -2,8 +2,10 @@ package fs2
 
 import cats.~>
 import cats.effect.IO
+import cats.implicits._
 import org.scalacheck.Gen
 import org.scalatest.Inside
+import scala.concurrent.duration._
 
 class StreamSpec extends Fs2Spec with Inside {
 
@@ -139,6 +141,52 @@ class StreamSpec extends Fs2Spec with Inside {
 
     "translate stack safety" in {
       Stream.repeatEval(IO(0)).translate(new (IO ~> IO) { def apply[X](x: IO[X]) = IO.suspend(x) }).take(1000000).run.unsafeRunSync()
+    }
+
+    "duration" in {
+      val delay = 200 millis
+
+      val blockingSleep = IO { Thread.sleep(delay.toMillis) }
+
+      val emitAndSleep = Stream.emit(()) ++ Stream.eval(blockingSleep)
+      val t = emitAndSleep zip Stream.duration[IO] drop 1 map { _._2 } runLog
+
+      (IO.shift >> t).unsafeToFuture collect {
+        case Vector(d) => assert(d.toMillis >= delay.toMillis - 5)
+      }
+    }
+
+    "every" in {
+      pending // Too finicky on Travis
+      type BD = (Boolean, FiniteDuration)
+      val durationSinceLastTrue: Pipe[Pure,BD,BD] = {
+        def go(lastTrue: FiniteDuration, s: Stream[Pure,BD]): Pull[Pure,BD,Unit] = {
+          s.pull.uncons1.flatMap {
+            case None => Pull.done
+            case Some((pair, tl)) =>
+              pair match {
+                case (true , d) => Pull.output1((true , d - lastTrue)) >> go(d,tl)
+                case (false, d) => Pull.output1((false, d - lastTrue)) >> go(lastTrue,tl)
+              }
+          }
+        }
+        s => go(0.seconds, s).stream
+      }
+
+      val delay = 20.millis
+      val draws = (600.millis / delay) min 50 // don't take forever
+
+      val durationsSinceSpike = Stream.every[IO](delay).
+        map(d => (d, System.nanoTime.nanos)).
+        take(draws.toInt).
+        through(durationSinceLastTrue)
+
+      (IO.shift >> durationsSinceSpike.runLog).unsafeToFuture().map { result =>
+        val (head :: tail) = result.toList
+        withClue("every always emits true first") { assert(head._1) }
+        withClue("true means the delay has passed: " + tail) { assert(tail.filter(_._1).map(_._2).forall { _ >= delay }) }
+        withClue("false means the delay has not passed: " + tail) { assert(tail.filterNot(_._1).map(_._2).forall { _ <= delay }) }
+      }
     }
   }
 }
